Reverse mixed-up headlines in readme
Just stumbled over this project and saw that there's a little mix up in the headings / paragraphs in the Readme. The WASM paragraph has the FaaS headline and the FaaS paragraph has the WASM headline.

This PR leaves the paragraph order as is: 1. FaaS 2. WASM and reverses the headlines so they match up.